Wording Tweaks
A friend pointed out that "'programmer' isn't used anymore". Google's Ngram shows that that may be a stretch. The term "programmer" by itself is still greatly in the lead for usage, but it's declining sharply. The terms "computer programmer" and "software engineer" are very closely matched, practically synonymous, but again "computer programmer" is on the decline, and "software engineer" is steadying out.